# dist/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Documentation GROUPIN ",
    "description": "Explication des apis du serveur de GROUPIN : https://github.com/CharlieMcFly/back-groupin",
    "termsOfService": ""
  },
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/users/groups" : {
      "post" : {
        "tags": [
          "Users"
        ],
        "description": "Rejoindre un groupe",
        "operationId": "for the Url",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "User Parametres",
            "in": "body",
            "description": "Groupe à ajouter à l'utilisateur",
            "required": true,
            "schema": {
              "$ref": "#/definitions/JoinGroup"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Renvoie l'utilisateur connecté, les notifications d'amis et de groupes",
            "schema": {
              "$ref": "#/definitions/ResponseUserNotifications"
            }
          }
        }
      }
    },
    "/users/{uid}/groups/{id}" : {
      "delete" : {
        "tags": [
          "Users"
        ],
        "description": "Quitter un groupe",
        "operationId": "for the Url",
        "produces": [
          "application/json"
        ],
        "parameters":[
          {
            "name": "uid",
            "in": "path",
            "description": "UID de l'utilisateur",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "ID du groupe",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Renvoie l'utilisateur et ses groupes",
            "schema":{
              "$ref":"#/definitions/ResponseUserGroups"
            }
          }
        }
      }
    },
    "/groups" : {
      "post": {
        "tags": [
          "Groups"
        ],
        "description": "Créer un groupe",
        "operationId": "for the Url",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Groupes Parametres",
            "in": "body",
            "description": "Groupe à ajouter à la base de données",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Group"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Renvoie l'utilisateur connecté et sa liste de groupe" ,
            "schema": {
              "$ref": "#/definitions/ResponseUserGroups"
            }
          }
        }
      }
    },
    "/groups/{uid}" : {
      "get": {
        "tags": [
          "Groups"
        ],
        "description": "Renvoie les groups d'un utilisateur",
        "operationId": "for the Url",
        "produces": [
          "application/json"
        ],
        "parameters":[
          {
            "name": "uid",
            "in": "path",
            "description": "UID de l'utilisateur",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Renvoie les groupes de l'utilisateur",
            "schema": {
              "$ref": "#/definitions/ResponseGroups"
            }
          }
        }
      }
    },
    "/groups/photo" : {
      "post": {
        "tags": [
          "Groups"
        ],
        "description": "Ajoute une photo au groupe",
        "operationId": "for the Url",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Groupes Parametres",
            "in": "body",
            "description": "Photo à ajouter au groupe dans la base de données",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Photo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Renvoie l'utilisateur connecté et sa liste de groupe" ,
            "schema": {
              "$ref": "#/definitions/ResponseGroup"
            }
          }
        }
      }
    },
    "/events/users/{uid}/groups/{key}" : {
      "get": {
        "tags": [
          "Events"
        ],
        "description": "Renvoie les events d'un groupe",
        "operationId": "for the Url",
        "produces": [
          "application/json"
        ],
        "parameters":[
          {
            "name": "uid",
            "in": "path",
            "description": "UID de l'utilisateur",
            "required": true,
            "type": "string"
          },
          {
            "name": "key",
            "in": "path",
            "description": "ID du groupe",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Renvoie les groupes de l'utilisateur",
            "schema": {
              "$ref": "#/definitions/ResponseUserEvents"
            }
          }
        }
      }
    },
...
    "notifications/groups" : {
      "post": {
        "tags": [
          "Notifications"
        ],
        "description": "Créer une notification pour une invitation dans un groupe",
        "operationId": "for the Url",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Notifications Parametres",
            "in": "body",
            "description": "Ajoute la notification de groupe à la base de données",
            "required": true,
            "schema": {
              "$ref": "#/definitions/JoinGroup"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "renvoie OK si aucun problème"
          }
        }
      }
    },
    "notifications/users/{uid}/groups/{idgroup}" : {
      "delete": {
        "tags": [
          "Notifications"
        ],
        "description": "Supprime une notification de groupe d'un user",
        "operationId": "for the Url",
        "produces": [
          "application/json"
        ],
        "parameters":[
          {
            "name": "uid",
            "in": "path",
            "description": "ID de l'utilisateur",
            "required": true,
            "type": "string"
          },
          {
            "name": "idgroup",
            "in": "path",
            "description": "ID du groupe",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Renvoie OK si tout est bon"
          }
        }
      }
    }
  },
  "definitions": {
    "User" : {
      "required" : [
        "uid", "email"
      ],
      "properties": {
        "uid": {
          "required":true,
          "type": "string",
          "example": "01544861548E26"
        },
        "email": {
          "required":true,
          "type":"string",
          "example": "test@test.be"
        },
        "displayName": {
          "required":true,
          "type":"string",
          "example": "Jean Valjean"
        },
        "photoURL": {
          "required":true,
          "type":"string",
          "example": "http://www.photo.com/unphoto.jpg"
        },
        "providerId": {
          "required":true,
          "type":"string",
          "example": "facebook.com"
        },
        "friends": {
          "required": false,
          "type": "object",
          "example": "{'15486151235': true }"
        },
        "events": {
          "required": false,
          "type": "object",
          "example": "{'15486151235': true }"
        },
        "groups": {
          "required": false,
          "type": "object",
          "example": "{'15486151235': true }"
        },
        "votes": {
          "required": false,
          "type": "object",
          "example": "{'15486151235': true }"
        }
      }
    },
    "Group" : {
      "required" : ["description", "photoURL", "nom"],
      "properties" : {
        "id" : {
          "required":true,
          "type": "string",
          "example": "01544861548E26"
        },
        "nom" : {
          "required":true,
          "type": "string",
          "example": "groupe de x"
        },
        "description" : {
          "required":true,
          "type": "string",
          "example": "description de x"
        },
        "membres" : {
          "required":true,
          "type": "object",
          "example": "{'10025148420' : true}"
        },
        "photoURL" : {
          "required":true,
          "type": "string",
          "example": "http://myphoto.jpg"
        },
        "votes" : {
          "required":true,
          "type": "object",
          "example": "{'10025148420' : true}"
        },
        "events" : {
          "required":true,
          "type": "object",
          "example": "{'10025148420' : true}"
        },
        "messages" : {
          "required":true,
          "type": "object",
          "example": "{'10025148420' : true}"
        },
        "photos" : {
          "required":true,
          "type": "object",
          "example": "{'10025148420' : true}"
        }
      }
    },
    "Event" : {
      "required" : ["nom", "photoURL", "description", "dateDebut", "dateFin"],
      "properties" : {
        "id" : {
          "required":true,
          "type": "string",
          "example": "01544861548E26"
        },
        "nom" : {
          "required":true,
          "type": "string",
          "example": "évènement de x"
        },
        "photoURL" : {
          "required":true,
          "type": "string",
          "example": "http://photo.jpg"
        },
        "description" : {
          "required":true,
          "type": "string",
          "example": "description de x"
        },
        "dateDebut" : {
          "required":true,
          "type": "string",
          "example": "125468221315"
        },
        "dateFin" : {
          "required":true,
          "type": "string",
          "example": "1235498451215"
        },
        "obj" : {
          "required":true,
          "type": "string",
          "example": "une bouteille"
        },
        "theme" : {
          "required":true,
          "type": "string",
          "example": "déguisé halloween"
        },
        "prix" : {
          "required":true,
          "type": "string",
          "example": "15"
        }
      }
    },
...
    "ResponseUserNotifications" : {
      "required" : ["user", "notifsAmis", "notifsGroupes"],
      "properties" : {
        "user" : {
          "required" : true,
          "$ref": "#/definitions/User"
        },
        "notifsAmis" : {
          "required" : true,
          "type" : "array",
          "items":{
            "$ref": "#/definitions/User"
          }
        },
        "notifsGroupes" : {
          "required" : true,
          "type" : "array",
          "items":{
            "$ref": "#/definitions/Group"
          }
        }
      }
    },
    "JoinGroup" : {
      "properties": {
        "idG": {
          "required":true,
          "type": "string",
          "example": "01544861548E26"
        },
        "uidR": {
          "required":true,
          "type":"string",
          "example": "01544861548E26"
        }
      }
    },
    "ResponseUserGroups" : {
      "required" : ["user","groups"],
      "properties" : {
        "user" : {
          "required" : true,
          "$ref": "#/definitions/User"
        },
        "groups" : {
          "required" : true,
          "type" : "array",
          "items":{
            "$ref": "#/definitions/Group"
          }
        }
      }
    },
    "ResponseGroups" : {
      "required" : ["groups"],
      "properties" : {
        "groups" : {
          "required" : true,
          "type" : "array",
          "items" : {
            "$ref": "#/definitions/Group"
          }
        }
      }
    },
    "Photo" : {
      "required" : ["idgroup", "url"],
      "properties" : {
        "idgroup" : {
          "required" : true,
          "type" : "string",
          "example" : "021568421"
        },
        "url" : {
          "required" : true,
          "type" : "string",
          "example" : "http://photo.jpg"
        }
      }
    },
    "ResponseGroup" : {
      "required" : ["group"],
      "properties" : {
        "group" : {
          "required" : true,
          "$ref": "#/definitions/Group"
        }
      }
    },
    "ResponseUserEvents" : {
      "required": ["user", "events"],
      "properties" : {
        "user" : {
          "required" : true,
          "$ref" : "#/definitions/User"
        },
        "events" : {
          "required" :  true,
          "type" : "array",
          "items"  : {
            "$ref" : "#/definitions/Event"
          }
        }
      }
    },
...
  }
}